Throw a new exception when we reach EOF.
The ReadableStreamingData's hasRemaining() method could return true even if the stream is empty, which will result in BufferUnderflowException thrown.
A new exception type BufferUnderflowException is added, so users can distinguish and catch the benign exception.

After talking to @jasperpotts, it became apparent this is the best approach to solve the iissue.

All the Codecs are modified to catch the exception.

Signed-off-by: Lubomir Litchev <lubomir@swirldslabs.com>