Some optimizations of Bytes.
This optimizations get us about 4-5% speadup. (Enables us to do optimizations later as well.

The Bytes object is now a ReadableSequentialData object as well.
Doing this optimization allows us to have private methods, instead of doing interface calls. That makes the Jit inline a lot of the code and gets us some perf.

Signed-off-by: Lubomir Litchev <lubomir@swirldslabs.com>