Register 'pbj' extension and task for all source sets
This allows users to generate code from 'proto' files in additional
source sets such as 'testFixtures'. The setup is done for all registered
source sets automatically. If source sets do not have 'pbj' sources,
the additional task is skipped with NO-SOURCE.

A special treatment is given to the 'test' source set as it contains
the generated tests for the 'main' source sets. For custom source
sets, tests are never generated. This special treatment and behavior
could be reconsidered in the future. For now, it preserves the current
behavior wrt. to 'main' and 'test'.

Signed-off-by: Jendrik Johannes <jendrik.johannes@gmail.com>