Fix poor hashing distribution in Bytes.
We have a really great hashing algorithm in the Hiero
Consensus Node in the NonCryptographicHashing file. This PR
brings it over here, but leaves it package private. A separate
PR should make it public and possibly move it somewhere else.
And a future PR should fix all other hashCode() implementations,
including for generated PBJ objects.

Signed-off-by: Richard Bair <rbair23@users.noreply.github.com>